# rest/build/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"NotFoundError": {
				"properties": {
					"message": {
						"type": "string",
						"enum": [
							"Not found"
						],
						"nullable": false
					}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"InternalError": {
				"properties": {
					"message": {
						"type": "string",
						"enum": [
							"Internal Server Error"
						],
						"nullable": false
					}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Response": {
				"description": "This Fetch API interface represents the response to a request.\n\n[MDN Reference](https://developer.mozilla.org/docs/Web/API/Response)",
				"properties": {},
				"type": "object",
				"additionalProperties": false
			},
			"ValidateErrorJSON": {
				"properties": {
					"message": {
						"type": "string",
						"enum": [
							"Validation failed"
						],
						"nullable": false
					},
					"details": {
						"type": "string"
					}
				},
				"required": [
					"message",
					"details"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Name": {
				"properties": {
					"firstName": {
						"type": "string"
					},
					"middleName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					}
				},
				"required": [
					"firstName",
					"lastName"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Pick_Doctor.Exclude_keyofDoctor.id-or-hireDate__": {
				"properties": {
					"name": {
						"$ref": "#/components/schemas/Name"
					},
					"specialization": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"specialization"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_Doctor.id-or-hireDate_": {
				"$ref": "#/components/schemas/Pick_Doctor.Exclude_keyofDoctor.id-or-hireDate__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"CreateDoctorRequest": {
				"$ref": "#/components/schemas/Omit_Doctor.id-or-hireDate_"
			},
			"Status": {
				"enum": [
					"scheduled",
					"completed"
				],
				"type": "string"
			},
			"Pick_Appointment.Exclude_keyofAppointment.id__": {
				"properties": {
					"patientId": {
						"type": "number",
						"format": "double"
					},
					"doctorId": {
						"type": "number",
						"format": "double"
					},
					"date": {
						"type": "string",
						"format": "date-time"
					},
					"reason": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/Status"
					}
				},
				"required": [
					"patientId",
					"doctorId",
					"date",
					"status"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_Appointment.id_": {
				"$ref": "#/components/schemas/Pick_Appointment.Exclude_keyofAppointment.id__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"CreateAppointmentRequest": {
				"$ref": "#/components/schemas/Omit_Appointment.id_"
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "vet-clinic-rest-server",
		"version": "1.0.0",
		"description": "Express.js based web service for managing veterinarian clinic",
		"contact": {
			"name": "Kondak Mariia"
		}
	},
	"paths": {
...
		"/api/doctors": {
			"get": {
				"operationId": "GetDoctors",
				"responses": {
					"200": {
						"description": "Fulfilled",
						"content": {
							"application/json": {
								"schema": {},
								"examples": {
									"Example 1": {
										"value": [
											{
												"id": 3,
												"name": {
													"firstName": "Jan",
													"lastName": "Pawel"
												},
												"specialization": "Internista",
												"hireDate": "2025-01-29T07:37:15.417Z",
												"email": "jan.pawel@clinic.com",
												"phone": "611892343"
											}
										]
									}
								}
							}
						}
					},
					"404": {
						"description": "Not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotFoundError"
								}
							}
						}
					},
					"500": {
						"description": "Unexpected error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalError"
								}
							}
						}
					}
				},
				"tags": [
					"Doctor"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateDoctor",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {},
								"examples": {
									"Example 1": {
										"value": {
											"id": 3,
											"name": {
												"firstName": "Jan",
												"lastName": "Pawel"
											},
											"specialization": "Internista",
											"hireDate": "2025-01-29T07:37:15.418Z",
											"email": "jan.pawel@clinic.com",
											"phone": "611892343"
										}
									}
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Response"
								}
							}
						},
						"headers": {
							"ValidateErrorJSON": {
								"schema": {
									"$ref": "#/components/schemas/ValidateErrorJSON"
								}
							}
						}
					},
					"500": {
						"description": "Unexpected error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalError"
								}
							}
						}
					}
				},
				"tags": [
					"Doctor"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateDoctorRequest"
							}
						}
					}
				}
			}
		},
		"/api/doctors/{id}": {
			"get": {
				"operationId": "GetDoctorById",
				"responses": {
					"200": {
						"description": "Fulfilled",
						"content": {
							"application/json": {
								"schema": {},
								"examples": {
									"Example 1": {
										"value": {
											"id": 3,
											"name": {
												"firstName": "Jan",
												"lastName": "Pawel"
											},
											"specialization": "Internista",
											"hireDate": "2025-01-29T07:37:15.418Z",
											"email": "jan.pawel@clinic.com",
											"phone": "611892343"
										}
									}
								}
							}
						}
					},
					"404": {
						"description": "Not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotFoundError"
								}
							}
						}
					},
					"500": {
						"description": "Unexpected error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalError"
								}
							}
						}
					}
				},
				"tags": [
					"Doctor"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateDoctor",
				"responses": {
					"200": {
						"description": "Updated",
						"content": {
							"application/json": {
								"schema": {},
								"examples": {
									"Example 1": {
										"value": {
											"id": 3,
											"name": {
												"firstName": "Jan",
												"lastName": "Pawel"
											},
											"specialization": "Internista",
											"hireDate": "2025-01-29T07:37:15.421Z",
											"email": "jan.pawel@clinic.com",
											"phone": "611892343"
										}
									}
								}
							}
						}
					},
					"404": {
						"description": "Not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotFoundError"
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Response"
								}
							}
						},
						"headers": {
							"ValidateErrorJSON": {
								"schema": {
									"$ref": "#/components/schemas/ValidateErrorJSON"
								}
							}
						}
					},
					"500": {
						"description": "Unexpected error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalError"
								}
							}
						}
					}
				},
				"tags": [
					"Doctor"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateDoctorRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteDoctor",
				"responses": {
					"200": {
						"description": "Deleted",
						"content": {
							"application/json": {
								"schema": {},
								"examples": {
									"Example 1": {}
								}
							}
						}
					},
					"404": {
						"description": "Not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotFoundError"
								}
							}
						}
					},
					"500": {
						"description": "Unexpected error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalError"
								}
							}
						}
					}
				},
				"tags": [
					"Doctor"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/appointments": {
			"get": {
				"operationId": "GetAppointments",
				"responses": {
					"200": {
						"description": "Fulfilled",
						"content": {
							"application/json": {
								"schema": {},
								"examples": {
									"Example 1": {
										"value": [
											{
												"id": 123,
												"patientId": 10,
												"doctorId": 3,
												"date": "2025-01-29T07:37:15.422Z",
												"status": "scheduled"
											}
										]
									}
								}
							}
						}
					},
					"404": {
						"description": "Not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotFoundError"
								}
							}
						}
					},
					"500": {
						"description": "Unexpected error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalError"
								}
							}
						}
					}
				},
				"tags": [
					"Appointment"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "doctorId",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "patientId",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateAppointment",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {},
								"examples": {
									"Example 1": {
										"value": {
											"id": 123,
											"patientId": 10,
											"doctorId": 3,
											"date": "2025-01-29T07:37:15.424Z",
											"status": "scheduled"
										}
									}
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Response"
								}
							}
						},
						"headers": {
							"ValidateErrorJSON": {
								"schema": {
									"$ref": "#/components/schemas/ValidateErrorJSON"
								}
							}
						}
					},
					"500": {
						"description": "Unexpected error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalError"
								}
							}
						}
					}
				},
				"tags": [
					"Appointment"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateAppointmentRequest"
							}
						}
					}
				}
			}
		},
		"/api/appointments/{id}": {
			"get": {
				"operationId": "GetAppoinmentById",
				"responses": {
					"200": {
						"description": "Fulfilled",
						"content": {
							"application/json": {
								"schema": {},
								"examples": {
									"Example 1": {
										"value": {
											"id": 123,
											"patientId": 10,
											"doctorId": 3,
											"date": "2025-01-29T07:37:15.424Z",
											"status": "scheduled"
										}
									}
								}
							}
						}
					},
					"404": {
						"description": "Not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotFoundError"
								}
							}
						}
					},
					"500": {
						"description": "Unexpected error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalError"
								}
							}
						}
					}
				},
				"tags": [
					"Appointment"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateAppointment",
				"responses": {
					"200": {
						"description": "Updated",
						"content": {
							"application/json": {
								"schema": {},
								"examples": {
									"Example 1": {
										"value": {
											"id": 123,
											"patientId": 10,
											"doctorId": 3,
											"date": "2025-01-29T07:37:15.428Z",
											"status": "scheduled"
										}
									}
								}
							}
						}
					},
					"404": {
						"description": "Not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotFoundError"
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Response"
								}
							}
						},
						"headers": {
							"ValidateErrorJSON": {
								"schema": {
									"$ref": "#/components/schemas/ValidateErrorJSON"
								}
							}
						}
					},
					"500": {
						"description": "Unexpected error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalError"
								}
							}
						}
					}
				},
				"tags": [
					"Appointment"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateAppointmentRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteAppointment",
				"responses": {
					"200": {
						"description": "Deleted",
						"content": {
							"application/json": {
								"schema": {},
								"examples": {
									"Example 1": {}
								}
							}
						}
					},
					"404": {
						"description": "Not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotFoundError"
								}
							}
						}
					},
					"500": {
						"description": "Unexpected error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalError"
								}
							}
						}
					}
				},
				"tags": [
					"Appointment"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "http://localhost:3000",
			"description": "Local development"
		}
	]
}